+++ Core/PKCE/HttpPkceRequestHandler.cs
@@ -8,7 +8,7 @@ using Unity.Cloud.Common;
 namespace Unity.Cloud.Identity
 {
     /// <summary>
-    /// This class handles all http requests required in the Proof Key Code Exchange authentication flow.
+    /// Handles all HTTP requests required in the Proof Key Code Exchange authentication flow.
     /// </summary>
     public class HttpPkceRequestHandler : IPkceRequestHandler
     {
@@ -16,9 +16,9 @@ namespace Unity.Cloud.Identity
         readonly IPkceConfigurationProvider m_PkceConfigurationProvider;
 
         /// <summary>
-        /// Creates a <see cref="HttpPkceRequestHandler"/> that handles all http requests required in the Proof Key Code Exchange authentication flow.
+        /// Creates a <see cref="HttpPkceRequestHandler"/> that handles all HTTP requests required in the Proof Key Code Exchange authentication flow.
         /// </summary>
-        /// <param name="httpClient">The <see cref="IHttpClient"/> instance required to make http requests.</param>
+        /// <param name="httpClient">The <see cref="IHttpClient"/> instance required to make HTTP requests.</param>
         /// <param name="pkceConfigurationProvider">The <see cref="IPkceConfigurationProvider"/> instance used to fetch the <see cref="PkceConfiguration"/> holding endpoints url.</param>
         public HttpPkceRequestHandler(IHttpClient httpClient, IPkceConfigurationProvider pkceConfigurationProvider)
         {
@@ -27,11 +27,11 @@ namespace Unity.Cloud.Identity
         }
 
         /// <summary>
-        /// Creates a Task that sends an http request to the <see cref="PkceConfiguration.TokenUrl"/> to exchange a code for a <see cref="DeviceToken"/>.
+        /// Creates a task that sends an HTTP request to the <see cref="PkceConfiguration.TokenUrl"/> to exchange a code for a <see cref="DeviceToken"/>.
         /// </summary>
         /// <param name="tokenEndPointParams">The application/x-www-form-urlencoded string value that holds all parameters required to reach the <see cref="PkceConfiguration.TokenUrl"/> endpoint.</param>
         /// <returns>
-        /// A Task that results in a <see cref="DeviceToken"/> when completed.
+        /// A task that results in a <see cref="DeviceToken"/> when completed.
         /// </returns>
         public async Task<DeviceToken> ExchangeCodeForDeviceTokenAsync(string tokenEndPointParams)
         {
@@ -42,12 +42,12 @@ namespace Unity.Cloud.Identity
         }
 
         /// <summary>
-        /// Creates a Task that sends an http request to the <see cref="PkceConfiguration.RefreshTokenUrl"/> to refresh the current <see cref="DeviceToken"/>.
+        /// Creates a task that sends an HTTP request to the <see cref="PkceConfiguration.RefreshTokenUrl"/> to refresh the current <see cref="DeviceToken"/>.
         /// </summary>
         /// <param name="tokenEndPointParams">The application/x-www-form-urlencoded string value that holds all parameters required to reach the <see cref="PkceConfiguration.RefreshTokenUrl"/> endpoint.</param>
         /// <param name="refreshToken">The refresh token string value to persist if no refresh token value is returned from <see cref="PkceConfiguration.RefreshTokenUrl"/> endpoint.</param>
         /// <returns>
-        /// A Task that results in a <see cref="DeviceToken"/> when completed.
+        /// A task that results in a <see cref="DeviceToken"/> when completed.
         /// </returns>
         public async Task<DeviceToken> RefreshTokenAsync(string tokenEndPointParams, string refreshToken)
         {
@@ -58,11 +58,11 @@ namespace Unity.Cloud.Identity
         }
 
         /// <summary>
-        /// Creates a Task that sends an http request to the <see cref="PkceConfiguration.LogoutUrl"/> to revoke the current <see cref="DeviceToken"/>.
+        /// Creates a Task that sends an HTTP request to the <see cref="PkceConfiguration.LogoutUrl"/> to revoke the current <see cref="DeviceToken"/>.
         /// </summary>
         /// <param name="revokeEndPointParams">The application/x-www-form-urlencoded string value that holds all parameters required to reach the <see cref="PkceConfiguration.LogoutUrl"/> endpoint.</param>
         /// <returns>
-        /// A Task.
+        /// A task.
         /// </returns>
         public async Task RevokeRefreshTokenAsync(string revokeEndPointParams)
         {