+++ Core/UserInformation/UserInfoClasses.cs
@@ -7,7 +7,7 @@ namespace Unity.Cloud.Identity
 {
 
     /// <summary>
-    /// This class holds the authenticated user information, like organizations the user belongs to (<see cref="OrganizationInfo"/>) and granted licenses (<see cref="LicenseInfo"/>).
+    /// Holds the authenticated user information, like organizations the user belongs to (<see cref="OrganizationInfo"/>) and granted licenses (<see cref="LicenseInfo"/>).
     /// </summary>
     public class UserInfo
     {