+++ Samples~/Documentation/Scripting/UserInfoProviderExample.cs
@@ -19,11 +19,11 @@ namespace Unity.Cloud.Identity.Documentation
 
     void CreateUserInfoProvider()
     {
-        var configuration = UnityRuntimeServiceHostConfigurationFactory.Create();
         var playerSettings = UnityCloudPlayerSettings.Instance;
         var httpClient = new UnityHttpClient();
+        var configuration = UnityRuntimeServiceHostConfigurationFactory.Create();
 
-        var compositeAuthenticatorSettings = new CompositeAuthenticatorSettingsBuilder(httpClient, PlatformSupportFactory.GetAuthenticationPlatformSupport())
+        var compositeAuthenticatorSettings = new CompositeAuthenticatorSettingsBuilder(httpClient, PlatformSupportFactory.GetAuthenticationPlatformSupport(), configuration)
             .AddDefaultBrowserAuthenticatedAccessTokenProvider()
             .AddDefaultPersonalAccessTokenProvider()
             .AddDefaultPkceAuthenticator(playerSettings, playerSettings)